requirements.txt
langchain
langchain_community
langchain_core
langchain_openai
streamlit
python-dotenv
openai